refactor: change int to Long

+++ bll/src/main/java/ru/hikkiyomi/service/CommonService.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface CommonService<T> {
-    Optional<T> findById(int id);
+    Optional<T> findById(Long id);
     void save(T obj);
     void delete(T obj);
     List<T> findAll();
