+++ bll/src/main/java/ru/hikkiyomi/service/KittenService.java
@@ -18,7 +18,7 @@ public class KittenService implements CommonService<Kitten> {
     }
 
     @Override
-    public Optional<Kitten> findById(int id) {
+    public Optional<Kitten> findById(Long id) {
         return Optional.ofNullable(dao.findById(id));
     }
