+++ dal/src/main/java/ru/hikkiyomi/dao/OwnerDao.java
@@ -7,5 +7,5 @@ import ru.hikkiyomi.model.Owner;
 @Repository
 public interface OwnerDao extends Dao<Owner>, JpaRepository<Owner, Integer> {
     @Override
-    Owner findById(int id);
+    Owner findById(Long id);
 }
